Add a global cache of open backend store handles
This allows concurrent operations to work even if different threads use
different handles.

Note: it is not clear who should own/write these caches: currently Irmin/core
relies on the backends to do this properly but maybe that should be managed
centrally and the core should have a cache of config/store handles somewhere.
Not totally clear we really want this, thought.